requirements.txt
pandas>=1.3.0
numpy>=1.20.0
openpyxl>=3.0.0
PyQt5>=5.15.0
xlrd>=2.0.1
Pillow>=9.0.0
pyinstaller>=5.0.0
jinja2>=3.0.0
sip>=6.0.0